refactor av from app
